.data to .text and ONE to ONEa
Changed constant to be stored in the .text section instead of
the .data section because the FIPS build can't handle the
data section.

Changed the name of the constant ONE to ONEa because the compiler
was complainng that ONE is a duplicate name. Not sure exactly why
but adding the suffix fixed the issue.