Add CI that uses M1 Mac ec2 Instance (#773)
The framework of this was taken from our original benchmarking
framework. The CDK stack creates a Codebuild, Cloudwatch, and
M1 ec2 instance resources, while we use the aws cli to spin up
an M1 ec2.
When a Codebuild job is kicked off, we use SSM (Systems Manager)
to send a command document to our M1 ec2. The results of this
can be observed in a Cloudwatch URL printed on the Console.
I envision this CI stack as something temporary we can use to
address our M1 CI gap. The end goal would be for us to migrate
to GIthub Action's M1 Mac workflows when they make that
available. This seems to be very near on the horizon (set for
Q3 2023).