Slightly optimize curve25519 field multiplication and squaring
Both bignum_mul_p25519 and bignum_sqr_p25519 in effect reduce
a double-length 2^256 * h + l to congruent 38 * h + l and then
further reduce based on a quotient estimate from the latter.
The new versions slightly reduce dependent latencies and/or
number of instructions by forming the quotient estimate before
the full computation of 38 * h + l and feeding the subsequent
correction in earlier as a result.