Add set up of OCSP_REQ_CTX with OCSP_REQUEST (#813)
Add functionality for setting up `OCSP_REQ_CTX` with an `OCSP_REQUEST`.
`OCSP_REQ_CTX` sets up the OCSP request to be sent over HTTP.

### Testing:
Generated OCSP Request DER files with the openssl ocsp tool, and added
generation method to the existing README. We mostly test if valid request files
are parsable, but we have an intentional failing case at the end.
Also added a test grabbing the internal contents of the |OCSP_REQ_CTX|, and
verify that both construction methods have the same output. There is also a
check that the constructed http header is there as expected, along with the
length of the OCSP Request contents.