Measure entire X25519 work for easy comparison with ECDH (#821)
For e.g. TLS, the entire flow is really "public key" generation and computation of the shared secret. For X25519, the AWS-LC functions doing that are: `X25519_public_from_private()` and `X25519()`. Where the latter also includes the randomness generation work.

But to make comparison with e.g. `bssl speed -filter ECDH` easy, we should output the performance of the entire flow as well.