Optimize low-UMULH-throughput ARM curve25519 field squaring
The "bignum_sqr_p25519" function and the basic 256-bit squaring
function "bignum_sqr_4_8" now break down the squaring computation into
a larger number of 32x32-bit multiplications, using Karatsuba
reduction at the top level, with only a few instances of UMULH. This
improves performance by about a factor of 1.4 on a typical targeted
microarchitecture where UMULH has low throughput. While it makes the
code slower on microarchitectures with high UMULH throughput, the
"_alt" forms of the functions are much faster there anyway and are
therefore to be preferred.