KEM: fix key retrieval when public/secret key is missing (#855)
Fixes key retrieval when public/secret key is missing. Specifically,
when `EVP_PKEY_get_raw_private_key` is called on a key that has
only the public key an error will be returned. The same when calling
`EVP_PKEY_get_raw_public_key` with a key that has only the
private key.

Got rid of the `has_secret_key field` in the `KEM_KEY` structure
because it is not necessary.

Made KEM `encapsulate` function a bit stricter with the function
arguments. Now, the arguments `ciphertext` and `shared_secret`
have to be either both NULL or both non-NULL. When both NULL
it's assumed the caller is performing a parameter length check.
When both non-NULL, the actual encapsulation is done. Otherwise,
the function fails.

---------

Co-authored-by: dkostic <dkostic@amazon.com>