Add AES-CCM/ECB/CTR/KW to 'test_fips' tool (#853) (#859)
AWS-LC currently only tests a few AES functions as a part of its tooling in test_fips.c. However, as a part of ATSEC's process, they wish to be able to step into functions for AES-CCM, ECB, CTR, AND KW with a debugger and thus have requested that we add them to our testing script.

In the interest of not having to repeat this patch for future validations, these functions were added to the test_fips tool the intention to bring these changes into the main, FIPS 1MU, and FIPS 2022 branches. Additionally, it would likely be prudent to bring the self_checks for the main and FIPS 2022 branches in line with these changes in another PR.