Add byte-level interfaces for X25519 functions
These provide alternative interfaces at the C level, with "_byte" in
their names to distinguish them, treating the arguments as arrays
of bytes (uint8_t) rather than of 64-bit words (uint64_t). This
better reflects how the X25519 function is generally specified and
used.

  void curve25519_x25519_byte(uint8_t res[static 32],uint8_t scalar[static 32],uint8_t point[static 32]);
  void curve25519_x25519_byte_alt(uint8_t res[static 32],uint8_t scalar[static 32],uint8_t point[static 32]);
  void curve25519_x25519base_byte(uint8_t res[static 32],uint8_t scalar[static 32]);
  void curve25519_x25519base_byte_alt(uint8_t res[static 32],uint8_t scalar[static 32]);

The underlying code is exactly the same in the x86 case, since
the platform is guaranteed to be little-endian, and the proofs
just rephrase the same results in terms of byte arrays.

The ARM functions are actually different code, using byte-level
loads and stores (ldrb, strb) at the beginning and end, and so
their proofs are also slightly different.