Convert X509_NAME_get_text_by_[NID|OBJ] to return UTF-8
Callers to these functions are usually using them to grab
subject name components and universally use the result as a C
string, whereas the OpenSSL versions return raw ASN1_STRING
bytes and ignore the encoding, which "usually works" in a
"hold my beer here are some bytes" sort of way until the
object is not encoded as you hoped.

Make this safer for the callers by making the returned
result be at least "text" and a C string. This converts
the ASN1_STRING bytes to UTF-8, and will introduce new
failure cases for this function if either memory allocation
fails for the UTF-8 conversion, or if the resulting UTF-8
contains a 0 codepoint and would produce an artificially
truncated C string. Additionally if the provided buffer
is not NULL but is too small to hold the output, we fail
rather than returning a truncated output.

Fixed: 436
Change-Id: I487c10a5ff5188e94df520ef4c8982e593c680d7
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/58445
Reviewed-by: David Benjamin <davidben@google.com>
Commit-Queue: Bob Beck <bbe@google.com>
(cherry picked from commit 3763efb56b5282cf92d71c259576352555c1a8f8)