Reference OPENSSL_ia32cap_P at Intel machine code level through a unique symbol instead of a common offset symbol avoiding add instruction (#862)
See #856 for background and description of issue resolved in this PR.

This is the second PR out of two PRs that bridge the performance gap.

See ticket for performance comparisons.

The first PR took care of the C-level. But the machine-optimised algorithm implementation sometimes directly dereference OPENSSL_ia32cap_P. These also need to be fixed-up. As before, we can't just add a call instruction to OPENSSL_ia32cap_get because it would compromise soundness/correctness.

Recall, that the issue was the add instruction. Instead of injecting that instruction, we instead do the following for each occurrence of OPENSSL_ia32cap_P discovered in the textual assembly:

* Define a new unique symbol. Uniquness is ensured using both the original (not true in all cases, but see "call outs") register name and a unique global counter. Bit redundant using both, but makes it easier to read IMO.
* Under the new unique symbol, copy the address of OPENSSL_ia32cap_P into the original register. This will probably spawn a relocation, but this is fine, because we put this new unique symbol outside the FIPS module scope in .text. Putting it here also means that we know the address relative to RIP.
* To jump back to the original execution point, we inject a "return" symbol/label where OPENSSL_ia32cap_P was discovered.