Let service indicator check for tls1.0/1.1 context (#927)
OpenSSL's way of doing this is checking the label and throwing an error when the
label "master secret"  is used. This is because "master secret" implies that the
PRF is used in a TLS1.0/1.1 context. In TLS1.2, the label is expected to be
"extended master secret". However, they only check if "master secret" is used
instead of checking if "extended master secret" is used, because the PRF can be
used for other purposes as well.
For our service indicator, we check if "extended master secret" is being used 
in TLS1.2. In TLS1.0/1.1, both the labels "master secret" and "extended master 
secret" can be used. We use the digests as an indicator of what TLS context
the PRF is being used in.