Ensure XTS mode API allows only re-initialising the tweak (#963)
XTS mode is sorta build to allow the tweak to change in a cheap way; requiring only one block cipher invocation on that tweak. The API supports this by setting the key argument to NULL but the "IV" argument (which is really the tweak when choosing the XTS implementation) to whatever tweak value desired.

Add unit tests to verify this and maintain the functionality in the future.