aesv8-armx.pl: Avoid buffer overread in AES-XTS decryption
Fixes: CVE-2023-1255

The buffer overread happens on decrypts of 4 mod 5 sizes.
Unless the memory just after the buffer is unmapped this is harmless.
This issue was reported on 27th February 2023 by Anton Romanov (Amazon).
The fix was developed by Nevine Ebeid (Amazon).

See OpenSSL master branch commit
https://github.com/openssl/openssl/commit/72dfe46550ee1f1bbfacd49f071419365bc23304
The fix is also available in commit bc2f61ad (for 3.1) and
commit 02ac9c94 (for 3.0).