Rename DILITHIUM3 OID/NID to DILITHIUM3_R3
This change qualifies the public DILITHIUM3 identifiers with a "round 3"
(`_R3`) suffix. Internal function, variable, etc. are left as-is because
they are not exposed to callers and can be updated if/when a materially
different round 4 comes down the pipeline.

Ideally, we'd be able to exclude this NID/OID entirely in non-dilithium
builds, but this is not currently possible with the current
`crypto/obj/` code. A next-best option would have been to put the
Dilithium3 NID into an "experimental" NID namespace _well_ above the
number of anticipated algorithms over the library's lifetime. I
prototyped this out with an NID of 10,001 but as it turns out the
tooling will expand NUM_NID and the `kObjects[]` in `obj_dat.h` and
backfill all ~9000 ivervening array values with NULL/0-valued entries.
This space tradeoff seemed unacceptable, so this implementation presents
the `DILITHIUM3_R3` NID in all builds.

The next commit will add a test to assert the behavior of trying to
specify `DILITHIUM3_R3` in non-`DILITHIUM_ENABLED=ON` builds.