add support for X509_get0_pubkey (#1000)
mySQL consumes the X509_get0_pubkey API when building with OpenSSL1.1.1 
and uses the X509_get_pubkey when built with OpenSSL1.0.2. X509_get0_pubkey
seems to have been added in OpenSSL1.1.1 and is very similar to X509_get_pubkey
(which exists in AWS-LC).

X509_get_pubkey attempts to decode the public key for certificate x. If
successful it returns the public key as an EVP_PKEY pointer with its reference
count incremented: this means the returned key must be freed up after use.
X509_get0_pubkey is similar except it does not increment the reference count of
the returned EVP_PKEY so it must not be freed up after use.