Add function for KEM key check
This change adds a public function EVP_PKEY_kem_check_key
that accepts an EVP_PKEY of type EVP_PKEY_KEM and returns 1
if the public and secret key from the given EVP_PKEY match.
Otherwise, it returns 0. The check is performed by doing
encapsulation/decapsulation and checking that the generated
shared secrets are equal.