Add word-level 59-fold iterated divstep
This uses a scaled delta, with the "d" argument corresponding to
2*delta for the delta in the original divstep and d assumed to be odd
(with the idea of starting the sequence of iterations at d = 1,
corresponding to delta = 1/2). Assuming d (and as usual f) to be
odd, the function performs 59 iterations of "divstep" and returns
a word-level 2x2 signed integer matrix indicating the updates to apply
to the initial f and g (which are expected to be usually larger than a
single word), as well as the updated d, which remains odd. While 59
looks a strange number, it corresponds well to the proven bounds on
divstep (e.g. 590 total iterations is natural for 256-bit inputs).