Fix OCSP integration tests to properly connect with endpoints (#1042)
The new integration test was failing, once Amazon Trust Services
reconfigured the behavior of their testing endpoints. The original
way the certificate chain had been configured on the endpoint had been
a bit off, and our tests had been depending on that irregular behavior.
The new behavior makes much more sense, but we've had to reconfigure
our tests to align with them.

1. I've removed the expired cert tests, since these don't fail in OCSP
anymore. They fail prematurely during the SSL connection due to the
expired cert. This new endpoint behavior is more aligned with
real-world behavior, but it doesn't belong within the realm of OCSP
testing anymore.
2. The endpoints had been sending the expected root certificate along
with the server's leaf/intermediate certificates over the wire. We had
been dependent on this behavior, but this was wrong in the first place.
We now store the intended root testing certificates in our local
testing environment, although this means we'll have to update the
certificates about 15 years down the line. I've added a warning print
function so we can be aware of this.
3. Some of these endpoints don't immediately present the intended
certificate chain to test. This is resolved by calling
SSL_set_tlsext_host_name, which forces the connection to use the leaf
certificate with the same host name as the endpoint.