Update sematests CI to run the ARM simulator
This patch updates the 'sematests' CI to run the simulator in parallel.
It is set to 30 minutes by default.

To achieve parallelism, a custom bash script is written and invoked from Makefile
instead of writing everything in Makefile.
This is because there is a challenge in relying on `make -j` to run simulators in parallel.
`make -j` assigns one subprocess for one target object, but in this case,
a simulator does not have any target object that it outputs.
To resolve this, we can create N artifical outputs, but then another challenge
is that in Makefile one cannot get the number of processes given to `-j`. :(
(see also: https://stackoverflow.com/q/5303553/1488216)

`make sematest` always runs the simulators using all cores in CPU.

Also, add '!/bin/bash' to the header of armsimulate.

This problem was found while Carl Kwan was running the simulator on his machine