Change the Jitter entropy recovery mechanism
Previously, we were calling the |jent_entropy_read_safe| Jitter function
to get entropy. This function has a built-in recover mechanism where
if the source fails to produce entropy then the function changes some
parameters and tries again. Due to FIPS rules, this is not possible
any more. We change the call to |jent_entropy_read| instead and handle
the potential failure in the AWS-LC part of the module by retrying
three times (note that in this case the same entropy source parameters
do not change).