Add edwards25519 combined fresh and base point scalar multiplication
The function edwards25519_scalarmuldouble[_alt] produces the sum of
two scalar products of points on the edwards25519 curve, i.e. a result
n * P + m * B where n and m are arbitrary scalar arguments, P is an
arbitrary point argument and B is the standard basepoint. It seems a
natural primitive for Ed25519 signature verification.