Save leaf certificate in SSL early to avoid losing external data (#1074)
This reverts some of the changes done in 
https://github.com/aws/aws-lc/commit/3a2b47ab5be5c75edacb8cdc246dc2dc8fb2c0cd. 
AWS-LC/BoringSSL also allows external data to be associated with the
X509 structures with the same APIs. However, our SSL connections don't
directly save the X509 certificate as a structure within `SSL_CTX`. We
"fold" the certificate and convert it into a binary blob to maintain
within `SSL_CTX`. The certificate is not parsed back into an `X509`
until the user asks for it.
This creates issues when the user attempts to associate data with the
certificate with `X509_set_ex_data` and passes the certificate into
`SSL_CTX`. The data usually associated is not ASN.1 code, thus
non-parsable, and the associated data gets lost during the translation
to pure bytes. This creates issues with codebases like nginx, who was
associating their own OCSP stapling data structures with X509 structures
 in SSL_CTX.

I've changed this to cache the leaf certificate early, so that the
external data is kept for that cert. The same behavior is kept for the
rest of the change. This keeps the changes needed contained and avoids an
entire restructure of the use of `CRYPTO_BUFFER` in `CERT`.
The leaf certificate could still be NULL if the user decides to pass in
a cert directly with ASN.1 encoding, so we keep the original behavior to
parse the cert on demand in this circumstance. This doesn't effect
`X509_set_ex_data` usage since external data can't be parsed.

This lets us pass more ssl_stapling in nginx. There are some more subtle
OCSP changes required, but I'll open another PR for that.