Align OCSP behavior with OpenSSL for nginx (#1077)
Asides from the behavior fixed in #1074, there were some slight OCSP
errors when building with nginx's OCSP stapling tests. There are still a
couple of tests failing due to us missing multiple certs support, but this
should resolve all other issues with us failing the test.

The NULL check in ocsp_find_signer_sk static function was also a bit
too pedantic, so I relaxed it. It's possible for the OCSP Basic Response
to be missing a cert stack, so we shouldn't be throwing an error on to
the stack if that happens. Also implemented functionality for the two
OCSP flags, OCSP_NOVERIFY and OCSP_TRUSTOTHER