Clean up handling of 32-bit operations in x86 model
This corrects the count masking for several 32-bit shift operations,
none of them actually used in s2n-bignum but nevertheless accepted by
the decoding setup. There are a few other tweaks as well that are not
functionally necessary but make the simulation smoother. In particular
BSWAP now just uses the standard HOL Light "word_bytereverse"
operation instead of a local variant.