add back CERT_PKEY structure and refactor pointers (#1086)
This doesn't actually implement multiple cert functionality, but sets up
the structures to do so. All certs currently use the first index
(SSL_PKEY_RSA) during the SSL connection.
A subsequent PR will be introduced to implement usage of the new
certificate slots. Overall functionality of the code shouldn't be
effected right now, this is just letting the code point to new pointers.