Add s2n-bignum-x86-sematests build project to CI
This patch adds a new build project to CI.

This is supposed to run a random simulator for x86 that @jargh has implemented.
I will make a follow-up PR that invokes it after this new CI is
incorporated smoothly.
At the moment, this build only runs an empty Makefile target `sematest`.