Allow `SSL_CTX` to use the new multiple certificate slots internally (#1100)
This builds upon the new framework in #1086 to use the new multiple
certificate slots available. The SSL connection should use the
corresponding index of the certificate and pkey type assigned. As of
now, the connection will directly use the last certificate that was
set for it's connection. This corresponds with our original behavior
before we reintroduced the multiple certificate slots framework,
where we only maintained one x509_leaf and one pkey per CERT.
A subsequent PR will be introduced to allow automated selection of the
correct certificate slots, based on the signature algorithm the
handshake has negotiated.

These functions change the certificate slot index:
* SSL_CTX_use_certificate_ASN1 / SSL_use_certificate_ASN1
SSL_CTX_use_certificate / SSL_use_certificate
(functions that call ssl_set_cert internally)
* SSL_use_*_PrivateKey_* (functions that call ssl_set_pkey internally)
* SSL_CTX_set_chain_and_key
The naming for the list of functions are pretty straightforward. Users
setting these functions would have expected SSL_CTX to use the new set
certificate/private key, so it's natural that we switch over to that
slot.

I've added a new test that sets up a connection with different 
certificate types and checks if the correct slot is being used.
ED25519 is incompatible with older versions of TLS, so those are
skipped.