Retrieve multiple certificate slot for TLS1.0/1.1 based on negotiated sigalgs (#1138)
This change is built on top of the multiple certificate slot change we
did for TLS1.2/1.3.
TLS1.0/1.1 only supports two different signature algorithms and is much
simpler. Either the RSA or ECDSA key type will indicate the use of it's
corresponding signature. This was what we were originally doing, so we
do the same for the private key slots, but with an additional check that
the cipher suite allows the key type.

I extended existing tests to use TLS1.0/1.1. More ssl runner tests will
be added in a subsequent commit.