fix build on gcc4.8.3 (#1156)
We recently had a team reach out to us about our broken build on gcc
4.8.3. This is due to a bug that we've previously encountered in this
commit: 65661035692572f9e399a162a4e908ad5b4bd509

We may have to add new CI to check for this, but it seems like this
specific version is hard to find. We already have CI for gcc 4.8.5.