Handle x30 target register in delocator for aarch64 to avoid clobbering the register (#1204)
When taking the commit: justsmth@3800f0f gcc 11.4 emitted the following in the routine ec_GFp_nistp224_felem_sqr():

```
adrp    x30, :got:__stack_chk_guard
```

This uncovered an unsupported case in the delocater.

When seeing the line above, the delocater resolved the relocation like this:

```
// WAS adrp    x30, :got:__stack_chk_guard
    sub sp, sp, 128
    stp x0, lr, [sp, #-16]!
    bl .Lboringssl_loadgot___stack_chk_guard
    mov x30, x0
    ldp x0, lr, [sp], #16
    add sp, sp, 128
```

But x30 is an alias for the aarch64 link register lr. Hence, mov x30, x0 would clobber x30 that is used as a general-purpose register in the un-modified routine; the routine pushes the state onto the stack and pops before returning making x30 available for scratch purposes. This caused a false-positive "stack smashing" event in the stack guard detector. The clobbered register x30 contained the stack guard value that is verified before returning from the routine. When the register got clobbered, the stack guard value got corrupted causing the validation to fail with very high probability.

Fix this by handling x30 as a special case in the delocater. If x30 is a target register, we assume that the routine use x30 as a (volatile) general-purpose register and will recover the link register value, using an appropriate method, before returning. This is really similar to how the x0 case is handled since x0 holds the return value from the subroutine and is not to be overwritten by the value saved on the stack..