Add back support for SSL_build_cert_chain (#1200)
This is a reimport of the code that was removed in boringssl@b2a9d6a.
Some of the function documentation was taken from OpenSSL's:
https://www.openssl.org/docs/man3.1/man3/SSL_CTX_build_cert_chain.html
This change ties into some of the code changes for multiple
certificate work that we had done recently. The code looks slightly
different from the original due to the C->C++ translation and
rearrangements for cleaner logic.