Add SDE+ASAN CI dimension (#1254)
Adding a CI dimension to support SDE+ASAN on Ubuntu 22.04. No memory
leaks were detected with the additional coverage, but some of these
tests hang and never exit when kicked off by the go test runner. The
current workaround is to set a 20 minute timer on the tests. Once the
timer is up and the process hasn't exited, we kill the process and
check stdout to see if the tests have actually passed or not. This
workaround is only applied to SDE as of now.
I've created CryptoAlg-2154 to help us investigate reasons this is
happening for SDE+ASAN, but it shouldn't block us from adding
new coverage for this.