Add CI for Windows MSVC2019, MSVC2022, and SDE 32/64-bit (#1228)
This creates a CI dimension for Windows 32 and 64 bit SDE. Windows
scripts were refactored to be more generic as well. We test Windows
SDE with the MSVC2019 build since Windows 11 has phased out some of
the more ancient processors.
The SHA256 dispatch test fails on the Debug build with newer
processors when building for 32-bit. I've tweaked the assumption for
the dispatch test and punted CryptoAlg-2137 to investigate the actual
reason. Curiously enough, this isn't an issue on the release build.
I also discovered how to properly build with MSVC2019 and MSVC2022
while working on this, so I added CI for the two compilers as well. I
tested all dimensions against appropriate instance sizes. SDE tests
take 90ish minutes to finish on 32 cores. We try to fit our CI
within 1 hour, so I went with the next largest instance (64 cores)
for the SDE simulator tests. Github provided Windows instances only
have 2 cores and aren't sufficient enough for our regular Windows
tests. The next larger instance was 8 cores, so I went with that.