Add a build option to the assembler to retain local symbols. (#1252)
Also, a test was added to ensure the module hash is the same with this option in a FIPS Release build. It uses a side effect (printout) of break-hash.go to get the module's hash value.
The test currently uses the FIPS Debug build since the Release build with gcc7 does not pass break-kat tests (higher versions of gcc pass). However, when modifying the script to omit break-kat tests and keep only break-hash test with a FIPS Release build, the gcc7 Release build passes.