Create dirs in cmake build dir with default permissions (#1269)
The downstream project, aws-iot-device-sdk-cpp-v2, fails to perform cmake configuration when sources are read-only. One of the issues is that aws-lc fails to perform cmake configuration for read-only sources.

By default, the file(COPY) cmake command preserves input file permissions. When source directory is read-only, this causes destination directory to be also read-only, which causes cmake configuration to fail when it tries to create a new file in the destination directory.
NO_SOURCE_PERMISSIONS option resolves the issue.

Co-authored-by: Igor Abdrakhimov <igorabd@amazon.com>