Add BFM, BIC, FCSEL, INS, SUB, TRN1, TRN2, USHR, ZIP2 to ARM model
BFM is the scalar instruction, analogous to the already-present UBFM
and SBFM. All the other new ones are vector instructions, INS meaning
both vector-to-vector "INS" and gpr-to-vector "INS (general)". The
existing vector MUL and SHL instruction models are also generalized
so that they cover more sizes. All newly modelled instructions are
validated by checks in the simulator.