Don't use expired certificates if possible. (#1282)
Some internal teams have ran into unsuspecting issues when migrating
over to AWS-LC. If an expired cert was placed before a possible valid
cert within the stack, AWS-LC would break and fail certificate
verification. This would not happen if the valid certificate was placed
first. The behavioral difference was pinned down to an older commit from
OpenSSL 1.0.2 -> OpenSSL 1.1.1: openssl/openssl@0930251, that we didn't
have support for. This reapplies that commit to get us more in line with
1.1.1 behavior.