Define program equivalence and rewrite `bignum_{mul,sqr}_8_16_neon` proofs using it
This patch rewrites the proofs of correctness of bignum_mul_8_16_neon and bignum_sqr_8_16_neon using program equivalence proof.
Their `*_SUBROUTINE_CORRECT` theorems (which are the most important theorems) remain unchanged. The only differences are that
their proofs rely on the program equivalence between their scalar versions and Neon versions.

To define program equivalence and prove related lemmas, this patch adds a bunch of definitions and lemmas in the following new files:
- `common/relational2.ml`: This file defines relational Hoare logic for machine code as a new predicate `ensures2`.
This internally uses a new predicate called `eventually_n`.
This file also defines a unary Hoare logic predicate `ensures_n` which uses `eventually_n` in contrast to `eventually` which is used by `ensures`.
Finally, this file proves several properties on `ensures2`, `ensures_n` and `ensures`.
- `arm/proofs/equiv.ml`: This file includes simulation tactics for a pair of programs. Also, this includes tactics and lemmas that enable conversion between Arm `ensures_n` and `ensures` because this is important to 'induce' the final specification from equivalence and original specification
- `tools/diff.py`: This is a Python script that prints the diff lines between two assembly files as an OCaml list of tuples. This is in turn used by `EQUIV_STEP_TAC`.

This patch only targets use cases in Arm, but `common/relational2.ml` is designed to be reusable for x86 assembly programs as well.